+++ Helper/QuestionHelper.php
@@ -11,7 +11,10 @@
 
 namespace Symfony\Component\Console\Helper;
 
+use Symfony\Component\Console\ConsoleEvents;
 use Symfony\Component\Console\Cursor;
+use Symfony\Component\Console\Event\QuestionAnsweredEvent;
+use Symfony\Component\Console\Exception\InvalidArgumentException;
 use Symfony\Component\Console\Exception\MissingInputException;
 use Symfony\Component\Console\Exception\RuntimeException;
 use Symfony\Component\Console\Formatter\OutputFormatter;
@@ -25,6 +28,8 @@ use Symfony\Component\Console\Question\ChoiceQuestion;
 use Symfony\Component\Console\Question\FileQuestion;
 use Symfony\Component\Console\Question\Question;
 use Symfony\Component\Console\Terminal;
+use Symfony\Component\Validator\Validation;
+use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 
 use function Symfony\Component\String\s;
 
@@ -37,6 +42,11 @@ class QuestionHelper extends Helper
 {
     private static bool $stty = true;
     private static bool $stdinIsInteractive;
+
+    public function __construct(
+        private ?EventDispatcherInterface $dispatcher = null,
+    ) {
+    }
 
     /**
      * Asks a question to the user.
@@ -59,7 +69,7 @@ class QuestionHelper extends Helper
         $inputStream ??= \STDIN;
 
         try {
-            if (!$question->getValidator()) {
+            if (!$question->getValidator() && !$question->getConstraints()) {
                 return $this->doAsk($inputStream, $output, $question);
             }
 
@@ -471,7 +481,17 @@ class QuestionHelper extends Helper
             }
 
             try {
-                return $question->getValidator()($interviewer());
+                $value = $interviewer();
+
+                if ($constraints = $question->getConstraints()) {
+                    $this->validateConstraints($value, $constraints);
+                }
+
+                if ($validator = $question->getValidator()) {
+                    return $validator($value);
+                }
+
+                return $value;
             } catch (RuntimeException $e) {
                 throw $e;
             } catch (\Exception $error) {
@@ -479,6 +499,27 @@ class QuestionHelper extends Helper
         }
 
         throw $error;
+    }
+
+    private function validateConstraints(mixed $value, array $constraints): void
+    {
+        if ($this->dispatcher) {
+            $event = new QuestionAnsweredEvent($value, $constraints);
+            $this->dispatcher->dispatch($event, ConsoleEvents::QUESTION_ANSWERED);
+
+            if ($event->hasViolations()) {
+                throw new InvalidArgumentException($event->getViolations()[0]);
+            }
+
+            return;
+        }
+
+        $validator = Validation::createValidator();
+        $violations = $validator->validate($value, $constraints);
+
+        if (\count($violations) > 0) {
+            throw new InvalidArgumentException($violations[0]->getMessage());
+        }
     }
 
     private function isInteractiveInput($inputStream): bool